Cleanup and add comment
